@ fix(sctelnewbma): ฟิกซ์ popup footer/เปิดซ้ำ + เพิ่มวันที่การเงิน/ชื่อ DB ใน topbar
Popup (pops/):
- PopOpen/PopCancelReason: แก้ปิดด้วย X/Esc แล้วเปิดซ้ำไม่ได้ —
  เปลี่ยน @bind-Visible → Visible+VisibleChanged แล้ว propagate กลับ Page
  (เดิม @bind อัปเดตแค่ local property, _openVisible/_cancelVisible ฝั่ง Page ค้าง true)
- ทุก popup มี footer + ปุ่มปิดอยู่ใน FooterContentTemplate
- ถอด fixed Height (PopOpen/PopValidateAML) → popup auto-size, footer pin ก้นเสมอ
  ไม่หลุดลอย; คุมความสูงตารางที่ .po-grid (height:300px) แทน

Topbar (MainLayout.razor):
- เพิ่มแถบสถานะ "วันที่การเงิน" + "ฐานข้อมูล" ข้างชื่อผู้ใช้ (เหมือน scCenter)
  ดึงจาก IScAppService.GetCenterInfoAsync() — แหล่งเดียวกับ scCenter
- chip ไม่มีพื้นหลัง/กรอบ (เอาออกตามที่ขอ)

Service (sc/):
- dbFactory.Database: คืนชื่อ DB ของ PG จาก connection string
- ScAppService.GetCenterInfoAsync: sysName = ชื่อ DB จริง (ไม่ใช่ sys_name ใน coop)
- CoopConfigDto.CountResign: int? → double? (count_resign เป็น double precision ใน PG)

@

+++ sc/SoAt.Application/ScTeller/sctelnewbma/SctelnewbmaDto.cs
@@ -23,7 +23,7 @@ public record ApplicationTypeDto(string Code, string? Name, decimal? Fee, string
 public record ConcernDto(string Code, string? RelatedNa);
 public record GroupPositionDto(string Code, string? Description, int? SortOrder);
 public record PositionDto(string Code, string? Name, int? SortOrder);
-public record CoopConfigDto(string CoopNo, int? CountResign, string? AutoApproveNewmem, string? MemTypeOngroup);
+public record CoopConfigDto(string CoopNo, double? CountResign, string? AutoApproveNewmem, string? MemTypeOngroup);  // count_resign = double precision ใน PG → ต้อง double? ไม่งั้น Dapper materialize ไม่ได้
 public record BankBranchDto(string Code, string? Name, string? BankId);   // cascade by BankId (เหมือน Districts ← Province)
 public record SalaryLevelDto(int Code, string? Name);                       // sc_mem_m_ucf_salary_level (ระดับ)
 public record SalaryRateDto(decimal Code, string? Name, int? LevelCode);    // sc_mem_m_ucf_salary_rate (ขั้น, cascade ← LevelCode)
